# product-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the CRUD NAAH PROJECT, a comprehensive system designed to manage shoe models and customer interactions.",
        "title": "CRUD NAAH PROJECT API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:50052",
    "basePath": "/api/v1",
    "paths": {
        "/customer/shoe-details": {
            "get": {
                "description": "Retrieve a list of all available shoe details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Fetch all shoe details for customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ShoeDetail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/customer/shoe-details/{id}": {
            "get": {
                "description": "Retrieve a specific shoe detail by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Fetch a shoe detail by ID for customers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shoe Detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShoeDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shoe-details": {
            "get": {
                "description": "Retrieve all shoe details including shoe_id, model_id, size, stock, name, and price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Get all shoe details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new shoe detail and model to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Create a new shoe detail and model",
                "parameters": [
                    {
                        "description": "Shoe Detail and Model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShoeDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ShoeDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shoe-details/{id}": {
            "get": {
                "description": "Retrieve a specific shoe detail by shoe_id including name and price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Get a shoe detail by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shoe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a shoe detail from the database using the shoe ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shoe-details"
                ],
                "summary": "Delete a shoe detail by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shoe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateShoeDetailRequest": {
            "type": "object",
            "properties": {
                "shoe_detail": {
                    "$ref": "#/definitions/models.ShoeDetail"
                },
                "shoe_model": {
                    "$ref": "#/definitions/models.ShoeModel"
                }
            }
        },
        "models.ShoeDetail": {
            "type": "object",
            "properties": {
                "model_id": {
                    "type": "integer"
                },
                "shoe_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "models.ShoeModel": {
            "type": "object",
            "properties": {
                "model_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        }
    }
}